Only optimize if both the left and right types are known because RHS can override the operation. Move binary work into a separate file.
